+++ src/main/java/capstone/lip/landinformationportal/service/SegmentOfStreetService.java
@@ -14,16 +14,28 @@ import capstone.lip.landinformationportal.repository.SegmentOfStreetRepository;
 import capstone.lip.landinformationportal.service.Interface.IFormedCoordinate;
 import capstone.lip.landinformationportal.service.Interface.IRealEstateAdjacentSegmentService;
 import capstone.lip.landinformationportal.service.Interface.ISegmentOfStreetService;
+import capstone.lip.landinformationportal.validation.SegmentOfStreetValidation;
 
 @Service
 public class SegmentOfStreetService implements ISegmentOfStreetService {
 
 	@Autowired
 	SegmentOfStreetRepository segmentOfStreetRepository;
-
+	
+	@Autowired
+	private IFormedCoordinate coordinateService;
+	
+	@Autowired
+	private IRealEstateAdjacentSegmentService adjService;
+	
 	@Override
 	public SegmentOfStreet save(SegmentOfStreet segmentOfStreet) {
 		try {
+			SegmentOfStreetValidation validate = new SegmentOfStreetValidation();
+			String error = validate.isValidSegment(segmentOfStreet);
+			if (!error.isEmpty()) {
+				throw new Exception(error);
+			}
 			return segmentOfStreetRepository.save(segmentOfStreet);
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -35,6 +47,11 @@ public class SegmentOfStreetService implements ISegmentOfStreetService {
 		try {
 			if (listSegment == null) throw new Exception("List segment is null");
 			if (listSegment.isEmpty()) throw new Exception("List segment is empty");
+			for (SegmentOfStreet element: listSegment) {
+				if (findById(element.getSegmentId())==null) {
+					throw new Exception("Segment not found");
+				}
+			}
 			List<FormedCoordinate> listCoordinate = listSegment.stream().map(x->x.getListFormedCoordinate()).flatMap(List::stream).collect(Collectors.toList());
 			coordinateService.delete(listCoordinate);
 			List<RealEstateAdjacentSegment> listAdj = listSegment.stream().map(x->x.getListRealEstateAdjacentSegment()).flatMap(List::stream).collect(Collectors.toList());
@@ -47,17 +64,15 @@ public class SegmentOfStreetService implements ISegmentOfStreetService {
 		}
 	}
 
-	@Autowired
-	private IFormedCoordinate coordinateService;
-	
-	@Autowired
-	private IRealEstateAdjacentSegmentService adjService;
-	
+
 	@Override
 	public boolean delete(SegmentOfStreet segmentOfStreet) {
 		
 		try {
 			if (segmentOfStreet == null) throw new Exception("Segment is null");
+			if (findById(segmentOfStreet.getSegmentId())==null) {
+				throw new Exception("Segment not found");
+			}
 			List<FormedCoordinate> listCoordinate = segmentOfStreet.getListFormedCoordinate();
 			coordinateService.delete(listCoordinate);
 			List<RealEstateAdjacentSegment> listAdj = segmentOfStreet.getListRealEstateAdjacentSegment();
@@ -73,6 +88,7 @@ public class SegmentOfStreetService implements ISegmentOfStreetService {
 	@Override
 	public SegmentOfStreet findById(Long id) {
 		try {
+			if (id==null) throw new Exception();
 			Optional<SegmentOfStreet> segment = segmentOfStreetRepository.findById(id);
 			if (segment.isPresent()) {
 				return segment.get();
