+++ src/main/java/capstone/lip/landinformationportal/service/StreetService.java
@@ -12,6 +12,7 @@ import capstone.lip.landinformationportal.entity.Street;
 import capstone.lip.landinformationportal.repository.StreetRepository;
 import capstone.lip.landinformationportal.service.Interface.ISegmentOfStreetService;
 import capstone.lip.landinformationportal.service.Interface.IStreetService;
+import capstone.lip.landinformationportal.validation.StreetValidation;
 
 @Service
 public class StreetService implements IStreetService{
@@ -24,6 +25,11 @@ public class StreetService implements IStreetService{
 	@Override
 	public Street save(Street street) {
 		try {
+			StreetValidation validate = new StreetValidation();
+			String error = validate.isValidStreet(street);
+			if (!error.isEmpty()) {
+				throw new Exception(error);
+			}
 			return streetRepository.save(street);
 		}catch(Exception e) {
 			e.printStackTrace();
@@ -36,6 +42,11 @@ public class StreetService implements IStreetService{
 		try {
 			if (listStreet == null) throw new Exception("List street is null");
 			if (listStreet.isEmpty()) throw new Exception("List street is empty");
+			for (Street element:listStreet) {
+				if (findById(element.getStreetId())==null) {
+					throw new Exception("Street not found");
+				}
+			}
 			List<SegmentOfStreet> listSegment = listStreet.stream().map(x->x.getListSegmentOfStreet()).flatMap(List::stream).collect(Collectors.toList());
 			segmentService.delete(listSegment);
 			
@@ -51,6 +62,9 @@ public class StreetService implements IStreetService{
 	public boolean delete(Street street) {
 		try {
 			if (street == null) throw new Exception("Street is null");
+			if (findById(street.getStreetId())==null) {
+				throw new Exception("Street not found");
+			}
 			List<SegmentOfStreet> listSegment = street.getListSegmentOfStreet();
 			segmentService.delete(listSegment);
 			
@@ -66,6 +80,7 @@ public class StreetService implements IStreetService{
 	@Override
 	public Street findById(Long id) {
 		try {
+			if (id==null) throw new Exception();
 			Optional<Street> street = streetRepository.findById(id);
 			if (street.isPresent()) {
 				return street.get();
